Socio-economic and demographic data of the Halifax Metropolitan Area for Dissemination Areas.

A simple features data frame with the boundary of Dissemination Areas in the Metropolitan Area. The boundary is unprojected (i.e., it is in latitude-longitude; CRS: WGS 84). The data frame contains a selection of socio-economic and demographic information for the Dissemination Areas.

Usage

data(census_data_da_hal)

Format

A simple features data frame with 601 row and 53 variables:

Type

Type of statistical area: "DA" for Dissemination Area

GeoUID

Unique identifier of the statistical area

region

Identifier of the region: "hal" for Halifax

Households

Total number of households in dissemination area in 2016

CD_UID

Unique identifier of the census CD

Dwellings

Total number of dwellings in dissemination area in 2016

Population

Total population in dissemination area in 2016

CSD_UID

Unique identifier of the census CSD

CT_UID

Unique identifier of the Census Tract

CMA_UID

Unique identifier of the census Census Metropolitan Area

tot_age_popn

CA 2016 Census; 100% data; Total - Age

age_0_to_14

CA 2016 Census; 100% data; Total - Age; 0 to 14 years

age_15_to_64

CA 2016 Census; 100% data; Total - Age; 15 to 64 years

age_65_plus

CA 2016 Census; 100% data; Total - Age; 65 years and over

tot_cens_fam_priv_hh

CA 2016 Census; 100% data; Census Families; Total number of census families in private households - 100% data

fem_parent

CA 2016 Census; 100% data; Census Families; Total number of census families in private households - 100% data; Total lone-parent families by sex of parent; Female parent

LICO_AT

CA 2016 Census; 100% data; Income; Low Income Measures; In low income based on the Low-income cut-offs, after tax (LICO-AT)

tot_LICO_status_2015

CA 2016 Census; 100% data; Income; Low Income Measures; Total - Low-income status in 2015 for the population in private households to whom low-income concepts are applicable - 100% data

knwl_off_lang

CA 2016 Census; 100% data; Language; Official Language; Total - Knowledge of official languages for the total population excluding institutional residents - 100% data

neither_EN_FR

CA 2016 Census; 100% data; Language; Official Language; Total - Knowledge of official languages for the total population excluding institutional residents - 100% data; Neither English nor French

Aborin_ID_priv_HH

CA 2016 Census; 25% Data; Aboriginal; Total - Aboriginal identity for the population in private households - 25% sample data

Aborin_ID

CA 2016 Census; 25% Data; Aboriginal; Total - Aboriginal identity for the population in private households - 25% sample data; Aboriginal identity

tot_priv_hh_hsing_suit

CA 2016 Census; 25% Data; Housing; Household Characteristics; Total - Private households by housing suitability - 25% sample data

not_suitable

CA 2016 Census; 25% Data; Housing; Household Characteristics; Total - Private households by housing suitability - 25% sample data; Not suitable

dwell_condtns

CA 2016 Census; 25% Data; Housing; Dwelling Characteristics; Total - Occupied private dwellings by dwelling condition - 25% sample data

maj_rep_needed

CA 2016 Census; 25% Data; Housing; Dwelling Characteristics; Total - Occupied private dwellings by dwelling condition - 25% sample data; Major repairs needed

tot_shelt_cost

CA 2016 Census; 25% Data; Housing; Shelter Cost; Total - Owner and tenant households with household total income greater than zero, in non-farm, non-reserve private dwellings by shelter-cost-to-income ratio - 25% sample data

shelt_gt30pc

CA 2016 Census; 25% Data; Housing; Shelter Cost; Total - Owner and tenant households with household total income greater than zero, in non-farm, non-reserve private dwellings by shelter-cost-to-income ratio - 25% sample data; Spending 30% or more of income on shelter costs

tot_tenant_HH

CA 2016 Census; 25% Data; Housing; Shelter Cost; Total - Tenant households in non-farm, non-reserve private dwellings - 25% sample data

tenant_HH_subsid

CA 2016 Census; 25% Data; Housing; Shelter Cost; Total - Tenant households in non-farm, non-reserve private dwellings - 25% sample data; % of tenant households in subsidized housing

tot_immig_stat

CA 2016 Census; 25% Data; Citizenship and Immigration; Total - Immigrant status and period of immigration for the population in private households - 25% sample data

immig_stat_last5

CA 2016 Census; 25% Data; Citizenship and Immigration; Total - Immigrant status and period of immigration for the population in private households - 25% sample data; Immigrants; 2011 to 2016

tot_adm_immig_popn_1980_to_2016

CA 2016 Census; 25% Data; Citizenship and Immigration; Total - Admission category and applicant type for the immigrant population in private households who landed between 1980 and 2016 - 25% sample data

refugees

CA 2016 Census; 25% Data; Citizenship and Immigration; Total - Admission category and applicant type for the immigrant population in private households who landed between 1980 and 2016 - 25% sample data; Refugees

visible_min

CA 2016 Census; 25% Data; Minority / Origin; Total - Visible minority for the population in private households - 25% sample data; Total visible minority population

tot_visible_min

CA 2016 Census; 25% Data; Minority / Origin; Total - Visible minority for the population in private households - 25% sample data

Black

CA 2016 Census; 25% Data; Minority / Origin; Total - Visible minority for the population in private households - 25% sample data; Total visible minority population; Black

tot_lbr_f_15_plus

CA 2016 Census; 25% Data; Work; Labour Force Status; Total - Population aged 15 years and over by Labour force status - 25% sample data

lbr_f_unemp

CA 2016 Census; 25% Data; Work; Labour Force Status; Total - Population aged 15 years and over by Labour force status - 25% sample data; In the labour force; Unemployed

tot_commuting_dur_lbr_f_15_plus

CA 2016 Census; 25% Data; Commute; Total - Commuting duration for the employed labour force aged 15 years and over in private households with a usual place of work or no fixed workplace address - 25% sample data

commuting_dur_60mins_plus

CA 2016 Census; 25% Data; Commute; Total - Commuting duration for the employed labour force aged 15 years and over in private households with a usual place of work or no fixed workplace address - 25% sample data; 60 minutes and over

from_mode_tran

CA 2016 Census; 25% Data; Commute; Total - Main mode of commuting for the employed labour force aged 15 years and over in private households with a usual place of work or no fixed workplace address - 25% sample data; Public transit

from_mode_walk

CA 2016 Census; 25% Data; Commute; Total - Main mode of commuting for the employed labour force aged 15 years and over in private households with a usual place of work or no fixed workplace address - 25% sample data; Walked

ave_HH_size

CA 2016 Census; 100% data; Households; Household size; Average household size

tot_degree_age_15_plus

CA 2016 Census; 25% Data; Education; Certificate/Degree; Total - Highest certificate, diploma or degree for the population aged 15 years and over in private households - 25% sample data

bachelor_or_higher

CA 2016 Census; 25% Data; Education; Certificate/Degree; Total - Highest certificate, diploma or degree for the population aged 15 years and over in private households - 25% sample data; Postsecondary certificate, diploma or degree; University certificate, diploma or degree at bachelor level or above

mhh_inc

CA 2016 Census; 100% data; Income; Households; Total - Income statistics in 2015 for private households by household size - 100% data; Median total income of households in 2015 ($)

from_mode_total

define_from_mode_total

from_mode_total

CA 2016 Census; 25% Data; Commute; Total - Main mode of commuting for the employed labour force aged 15 years and over in private households with a usual place of work or no fixed workplace address - 25% sample data

from_mode_pass

CA 2016 Census; 25% Data; Commute; Total - Main mode of commuting for the employed labour force aged 15 years and over in private households with a usual place of work or no fixed workplace address - 25% sample data; Car, truck, van - as a passenger

study_region_name

Name of the Census Metropolitan Area (CMA)

Total_Emp

define_Total_Emp

geometry

Simple features geometry

Source

Population: Canada Census 2016. See: https://mountainmath.github.io/cancensus/index.html

Examples

 data(census_data_da_hal)