Socio-economic and demographic data of the Vancouver Metropolitan Area for Dissemination Areas.

A simple features data frame with the boundary of Dissemination Areas in the Metropolitan Area. The boundary is unprojected (i.e., it is in latitude-longitude; CRS: WGS 84). The data frame contains a selection of socio-economic and demographic information for the Dissemination Areas.

Usage

data(census_data_da_van)

Format

A simple features data frame with 3,715 row and 53 variables:

Type

Type of statistical area: "DA" for Dissemination Area

GeoUID

Unique identifier of the statistical area

region

Identifier of the region: "van" for Vancouver

Households

Total number of households in dissemination area in 2016

CD_UID

Unique identifier of the census CD

Dwellings

Total number of dwellings in dissemination area in 2016

Population

Total population in dissemination area in 2016

CSD_UID

Unique identifier of the census CSD

CT_UID

Unique identifier of the Census Tract

CMA_UID

Unique identifier of the census Census Metropolitan Area

tot_age_popn

Total age population

age_0_to_14

Total population aged 0-14 years old in dissemination area in 2016

age_15_to_64

Total population aged 15-64 years old in dissemination area in 2016

age_65_plus

Total population aged 65 years and older in dissemination area in 2016

tot_cens_fam_priv_hh

Total number of census families in private households in dissemination area in 2016

fem_parent

define_fem_parent

LICO_AT

define_lico_at

tot_LICO_status_2015

define_tot_LICO

knwl_off_lang

define_knwl_off_lang

neither_EN_FR

define_neither_EN_FR

Aborin_ID_priv_HH

define_Aborin_ID_priv_HH

Aborin_ID

define_aborivin_ID

tot_priv_hh_hsing_suit

define_tot_priv_hh_hsing_suit

not_suitable

define_not_suitable

dwell_condtns

define_dwell_condtns

maj_rep_needed

define_maj_rep_needed

tot_shelt_cost

define_tot_shelt_cost

shelt_gt30pc

define_shelt_gt30pc

tot_tenant_HH

define_tot_tenant_HH

tenant_HH_subsid

define_tenant_HH_subsid

tot_immig_stat

define_tot_immig_stat

immig_stat_last5

define_immig_stat_last5

tot_adm_immig_popn_1980_to_2016

define_tot_adm_immig_popn_1980_to_2016

refugees

define_refugees

visible_min

define_visible_min

tot_visible_min

define_tot_visible_min

Black

define_Black

tot_lbr_f_15_plus

define_tot_lbr_f_15_plus

lbr_f_unemp

define_lbr_f_unemp

tot_commuting_dur_lbr_f_15_plus

define_tot_commuting_dur_lbr_f_15_plus

commuting_dur_60mins_plus

define_commuting_dur_60mins_plus

from_mode_tran

define_from_mode_tran

from_mode_walk

define_from_mode_walk

ave_HH_size

define_ave_HH_size

tot_degree_age_15_plus

define_tot_degree_age_15_plus

bachelor_or_higher

define_bachelor_or_higher

mhh_inc

define_mhh_inc

from_mode_total

define_from_mode_total

from_mode_car

define_from_mode_car

from_mode_pass

define_from_mode_pass

study_region_name

Name of the Census Metropolitan Area (CMA)

Total_Emp

define_Total_Emp

geometry

Simple features geometry

Source

Population: Canada Census 2016. See: https://mountainmath.github.io/cancensus/index.html

Examples

 data(census_data_da_van)