Population statistics for selected metropolitan areas in Canada.

A data frame with population for selected Census Metropolitan Areas in Canada for 2021.

Usage

data(population_statistics)

Format

A data frame with 14 rows and 5 variables:

size_rank

Rank of the metropolitan area according to population

year

The date of the data

cma_name

Name of the Census Metropolitan Area

province

Name of the province that is home to the Census Metropolitan Area

population

Population of the Census Metropolitan Area

Source

Statistics Canada.

Examples

 data(population_statistics)
 summary(population_statistics)
#>    size_rank      year                              cma_name
#>  1      :1   Min.   :2021   Abbotsford-Mission          :1  
#>  2      :1   1st Qu.:2021   Calgary                     :1  
#>  3      :1   Median :2021   Edmonton                    :1  
#>  4      :1   Mean   :2021   Halifax                     :1  
#>  5      :1   3rd Qu.:2021   Hamilton                    :1  
#>  6      :1   Max.   :2021   Kitchener-Cambridge-Waterloo:1  
#>  (Other):8                  (Other)                     :8  
#>                province   population     
#>  Alberta           :2   Min.   : 195726  
#>  British Columbia  :2   1st Qu.: 551625  
#>  Manitoba          :1   Median : 836994  
#>  Nova Scotia       :1   Mean   :1584309  
#>  Ontario           :5   3rd Qu.:1486682  
#>  Ontario and Quebec:1   Max.   :6202225  
#>  Quebec            :2