National and regional transit revenue and ridership in Canada.

A data frame with monthly ridership and revenue statistics for urban transit agencies, from April, 2017 to August, 2024.

Usage

data(transit_statistics)

Format

A data frame with 368 rows and 4 variables:

GEO

Names of four geographies as a factor: Canada, Atlantic, Quebec and Ontario, Prairies, British Columbia and Territories

REF_DATE

The date of the data

Revenue

Revenue of transit agencies, before subsidies, in dollars

Ridership

Total number of passengers

Source

Statistics Canada. Table 23-10-0251-01 Passenger bus and urban transit statistics, by the North American Industry Classification System (NAICS) (x 1,000,000) (https://doi.org/10.25318/2310025101-eng)

Examples

 data(transit_statistics)
 summary(transit_statistics)
#>                                          GEO        REF_DATE         
#>  Atlantic                                  :92   Min.   :2017-01-01  
#>  Canada                                    :92   1st Qu.:2018-11-23  
#>  Prairies, British Columbia and Territories:92   Median :2020-10-16  
#>  Quebec and Ontario                        :92   Mean   :2020-10-15  
#>                                                  3rd Qu.:2022-09-08  
#>                                                  Max.   :2024-08-01  
#>                                                                      
#>     Revenue            Ridership        
#>  Min.   :  1100000   Min.   :   500000  
#>  1st Qu.: 31150000   1st Qu.: 15350000  
#>  Median : 94650000   Median : 46400000  
#>  Mean   :135931928   Mean   : 59645482  
#>  3rd Qu.:236600000   3rd Qu.:100650000  
#>  Max.   :369000000   Max.   :169200000  
#>  NA's   :36          NA's   :36