Travel time to groceries for Edmonton Metropolitan Area.

A data frame with origin-destination pairs, travel time by transit, and population by Census Dissemination Area of origin. Dissemination Areas (DAs) are the smallest publicly available geography provided by Statistics Canada. Each destination is a single grocery opportunity, DEFINE THE SELECTION CRITERIA. Travel times are for a representative day in two different years: 2019 (pre-covid measures) and 2023 (post-covid measures) from the centroid of the origin DA to the address of the grocery. Times are calculated as the median of all times within a 15 minutes window of the departure.

Usage

data(travel_matrix_grc_edm)

Format

A data frame with 324,030 rows and 8 variables:

from_id

Unique identifier of the Dissemination Area of the zone of origin

to_id

Unique identifier of the Dissemination Area of the zone of destination

travel_time_p50

Median travel time in minutes by transit within a 15 minute window of the time of departure of 8:00 am

region

Identifier of the region: "edm" for Edmonton

start_datetime

Date and time of departure for a trip that originates at the Dissemination Area of origin

start_year

Year for which travel time was calculated

study_region_name

Name of the Census Metropolitan Area (CMA)

Population

Total population in the Dissemination Area of origin (2016 Census)

Source

Population: Canada Census 2016. See: https://mountainmath.github.io/cancensus/index.html

Groceries: DMTI Points of Interest Data

Travel times: calculated based on General Transit Feed Standard and OpenStreetMaps data

Examples

 data(travel_matrix_grc_edm)
 summary(travel_matrix_grc_edm$travel_time_p50)
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    0.00   51.00   67.00   68.07   84.00  120.00